requirements.txt
flask
pyserial
pysqlite3
pynmea2
smbus2
gpiozero
